Exception Handling

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -1,17 +1,20 @@
 package com.devsuperior.dscommerce.services;
 
-import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.devsuperior.dscommerce.dto.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DatabaseException;
 import com.devsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -48,17 +51,30 @@ public class ProductService {
 	// PUT
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) { // Id + o corpo (body)
-		Product entity = repository.getReferenceById(id); 
-		// Não vai no banco de dados (é monitorado pala JPA), é uma referência
-		copyDtoToEntity(dto, entity);
-		entity = repository.save(entity); // É assim que se salva no BD
-		return new ProductDTO(entity); // Retorno em formato DTO
+		try {
+			Product entity = repository.getReferenceById(id); 
+			// Não vai no banco de dados (é monitorado pala JPA), é uma referência
+			copyDtoToEntity(dto, entity);
+			entity = repository.save(entity); // É assim que se salva no BD
+			return new ProductDTO(entity); // Retorno em formato DTO
+		} catch(EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		
+		
 	}
 	
 	// DELETE
-	@Transactional // Assim digo que estou apenas lendo(fica mais rápido)
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if(!repository.existsById(id)) 
+			throw new ResourceNotFoundException("Recurso não encontrodo");
+		try {
+			repository.deleteById(id);
+		}
+		catch(DataIntegrityViolationException e) {
+	        throw new DatabaseException("Falha de integridade referencial");
+	   	}
 	}
 
 	private void copyDtoToEntity(ProductDTO dto, Product entity) {
